move packet parsing to root lib
